// src/cast.types.d.ts
export interface CastMediaInfo {
    contentId: string;
    contentType: string;
    streamType?: StreamType;
    metadata?: CastMetadata;
    textTracks?: CastTextTrack[];
    duration?: number;
    customData?: any;
    startAbsoluteTime?: number;
    textTrackStyle?: CastTextTrackStyle;
}
// ...
export interface CastTextTrack {
    id?: number;
    src: string;
    contentType: string;
    name: string;
    language: string;
}
export interface CastTextTrackStyle {
    backgroundColor?: string;
    customData?: any;
    edgeColor?: any;
    edgeType?: any;
    fontFamily?: any;
    fontGenericFamily?: any;
    fontScale?: any;
    fontStyle?: any;
    foregroundColor?: any;
    windowColor?: any;
    windowRoundedCornerRadius?: any;
    windowType?: any;
}
export interface CastMetadata {
    metadataType: MetadataType;
    images?: {
        url: string;
        width: number;
        height: number;
    }[];
    creationDate?: string;
    releaseDate?: string;
    broadcastDate?: string;
    title?: string;
    subtitle?: string;
    description?: string;
    artist?: string;
    albumArtist?: string;
    albumTitle?: string;
    composer?: string;
    discNumber?: string;
    trackNumber?: string;
    seasonNumber?: string;
    episodeNumber?: string;
    seriesTitle?: string;
    studio?: string;
    width?: string;
    height?: string;
    locationName?: string;
    locationLatitude?: string;
    locationLongitude?: string;
}
export declare enum PlayerState {
    UNKNOWN = "UNKNOWN",
    IDLE = "IDLE",
    PLAYING = "PLAYING",
    PAUSED = "PAUSED",
    BUFFERING = "BUFFERING",
    LOADING = "LOADING"
}
export declare enum ResumeState {
    UNCHANGED = "UNCHANGED",
    PLAY = "PLAY",
    PAUSE = "PAUSE"
}
export declare enum CastEvent {
    onProviderAdded = "onProviderAdded",
    onProviderChanged = "onProviderChanged",
    onProviderRemoved = "onProviderRemoved",
    onRouteAdded = "onRouteAdded",
    onRoutePresentationDisplayChanged = "onRoutePresentationDisplayChanged",
    onRouteRemoved = "onRouteRemoved",
    onRouteSelected = "onRouteSelected",
    onRouteUnselected = "onRouteUnselected",
    onSessionEnded = "onSessionEnded",
    onSessionEnding = "onSessionEnding",
    onSessionResumed = "onSessionResumed",
    onSessionResuming = "onSessionResuming",
    onSessionStarted = "onSessionStarted",
    onSessionStartFailed = "onSessionStartFailed",
    onSessionStarting = "onSessionStarting",
    onSessionSuspended = "onSessionSuspended",
    onDeviceVolumeChanged = "onDeviceVolumeChanged",
    onDeviceChanged = "onDeviceChanged",
    onMediaStatusChanged = "onMediaStatusChanged",
    mediaQueueWillChange = "mediaQueueWillChange",
    itemsReloaded = "itemsReloaded",
    itemsInsertedInRange = "itemsInsertedInRange",
    itemsUpdatedAtIndexes = "itemsUpdatedAtIndexes",
    itemsRemovedAtIndexes = "itemsRemovedAtIndexes",
    mediaQueueChanged = "mediaQueueChanged"
}
export declare enum MetadataType {
    MOVIE = "MOVIE",
    TV_SHOW = "TV_SHOW",
    MUSIC_TRACK = "MUSIC_TRACK",
    PHOTO = "PHOTO",
    USER = "USER",
    AUDIO_BOOK_CHAPTER = "AUDIO_BOOK_CHAPTER",
    GENERIC = "GENERIC"
}
export declare enum StreamType {
    BUFFERED = "BUFFERED",
    LIVE = "LIVE",
    UNKNOWN = "UNKNOWN",
    NONE = "NONE"
}
// ...
export interface QueueItem {
    mediaInformation: CastMediaInfo;
    itemID: number;
    autoplay: boolean;
    startTime: number;
    playbackDuration: number;
    preloadTime: number;
    activeTrackIds?: number[] | null;
    customData?: any;
}